added fetch calls to API utility
