+++ src/App.js
@@ -6,6 +6,7 @@ import Admin from './pages/Home/Admin'
 import SignIn from './pages/Home/SignIn'
 import Content from './pages/Home/Content'
 import {BrowserRouter as Router, Switch, Route, Redirect} from 'react-router-dom'
+import API from './util/api'
 class App extends React.Component {
   constructor(){
     super()
@@ -24,50 +25,14 @@ class App extends React.Component {
       mode: 'xml'
     }
     this.getAuthToken = this.getAuthToken.bind(this);
-    this.getTopics = this.getTopics.bind(this);
-    this.getModules = this.getModules.bind(this);
-    this.getLessons = this.getLessons.bind(this);
-    this.getQuestions = this.getQuestions.bind(this);
-    this.getResources = this.getResources.bind(this);
+    API.getTopics = API.getTopics.bind(this);
+    API.getModules = API.getModules.bind(this);
+    API.getLessons = API.getLessons.bind(this);
+    API.getQuestions = API.getQuestions.bind(this);
+    API.getResources = API.getResources.bind(this);
   }
   
-  async getTopics(){
-    let response = await fetch('/topics');
-    if(response.status === 200){
-      let data = await response.json();
-      this.setState({topics:data})
-    }
-  }
-  async getModules(){
-    let response = await fetch('/code_modules');
-    let data = await response.json();
-    if (response.status === 200) {
-      this.setState({modules:data})
-    }
-  }
-  async getLessons(){
-    let response = await fetch('/lessons')
-    let data = await response.json();
-    if(response.status === 200){
-      this.setState({lessons:data})
-    }
-  }
-  async getQuestions(){
-    let response = await fetch('/questions')
-    let data = await response.json();
-    if(response.status === 200) {
-      this.setState({questions:data}
-      )
-    }
-  }
-  async getResources(){
-    
-    let response = await fetch('/resources')
-    let data = await response.json();
-    if (response.status === 200) {
-      this.setState({resources:data})
-    }
-  }
+  
   getAuthToken = () => {
     if (localStorage.getItem('authToken') && localStorage.getItem('authToken') !== null) {
         let token = localStorage.getItem('authToken')
@@ -134,11 +99,11 @@ class App extends React.Component {
     })
   }
   componentDidMount(){
-    this.getTopics();
-    this.getModules();
-    this.getLessons();
-    this.getQuestions();
-    this.getResources();
+    API.getTopics();
+    API.getModules();
+    API.getLessons();
+    API.getQuestions();
+    API.getResources();
     this.setState({ mode: 'xml'});
     this.getAuthToken(this.findModule);
   }
